refactor(components): split src/components/ into domain subdirectories
Organizes components into:
- workspace/  TabBar, CardLayout, CardTree, WorkspaceOverlay, SavedConfigMenu
- card/       Card, PluginHost, IframePluginHost, EmptyState, PluginBrowser

Updates all import paths. No logic changes.

Closes #118
