+++ tsconfig.json
@@ -25,8 +25,6 @@
       "@origin/hello/*": ["./packages/hello/src/*"],
       "@origin/notepad": ["./packages/notepad/src"],
       "@origin/notepad/*": ["./packages/notepad/src/*"],
-      "@origin/sdk": ["./packages/sdk/src/index.ts"],
-      "@origin/sdk/*": ["./packages/sdk/src/*"]
     }
   },
   "include": ["src"],
